requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
idna==3.10
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-minio-storage==0.5.7
gunicorn==20.1.0
dj-database-url==2.1.0
